requirements.txt
python-pptx>=0.6.21
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
comtypes>=1.2.0
pillow>=10.0.0
